+++ core/src/com/phoenixkahlo/hellcraft/carbonite/package.scala
@@ -2,6 +2,6 @@ package com.phoenixkahlo.hellcraft
 
 package object carbonite {
 
-  type NodeTypeID = Byte
+  type NodeTypeID = Short
 
 }
